# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BUCO TV API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Channel {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  logo?: string | null;
  stream: string;
  featured?: boolean;
  /** @nullable */
  description?: string | null;
}

export type ChannelStatsCategoryCounts = {[key: string]: number};

export interface ChannelStats {
  total: number;
  featured: number;
  categoryCounts: ChannelStatsCategoryCounts;
}

export interface ErrorResponse {
  error: string;
}

export type ListChannelsParams = {
/**
 * Filter by category
 */
category?: string;
/**
 * Filter by featured status
 */
featured?: boolean;
};