# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BUCO TV API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  logo?: string | null;
  stream: string;
  featured?: boolean;
  /** @nullable */
  description?: string | null;
}