# lib/api-zod/src/generated/types/channelStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BUCO TV API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelStatsCategoryCounts } from './channelStatsCategoryCounts';

export interface ChannelStats {
  total: number;
  featured: number;
  categoryCounts: ChannelStatsCategoryCounts;
}